+++ dejafu/Test/DejaFu/Conc/Internal/Program.hs
@@ -5,6 +5,7 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 
 -- GHC doesn't need this to compile the module, but stylish-haskell
 -- does to format it.